Extract lean kirra-fleet-types crate and FleetTrustStore seam
